+++ employees.csv
@@ -1,3 +1,3 @@
-first_name, last_name, employee_id, DOE, salary, department
+firstName, lastName, startDate, salary, department, empID
 pablo, castro, 123, 1/23/2023, 50000, IT
 peter, pan, 321, 4/21/2018, 73000, HR
